Split GadgetsStreamController into GadgetStreamController (which handles the socket) and GadgetStreamInterface (which stores and communicates with the Gadgets on the stream
